requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
PyYAML==6.0.1
aiofiles
python-multipart
requests
